[CAMEL-13564] Move more methods to ErrorHandlerReifier
